Initialize program options with a vector
Boost (1.68 at least) assumes for the int+pointer constructor that there
is one element in the array (the program name) and starts copying
at argv+1, up to argv+argc, so from 1 to 0 (nullptr).

   25 basic_command_line_parser(int argc, const charT* const argv[])
   26 : detail::cmdline(
-> 27    to_internal(std::vector<std::basic_string<charT> >(argv+1, argv+argc))),
   28    m_desc()

This by itself is bad, but somewhere on the stack it would also
seem to substract end-start and cast to an unsigned. Since the
result is compared to a max acceptable value somewhere, it raises
an exception.